+++ yash-syntax/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ????-??-??
 
+### Added
+
+- `Lexer::pending`
+- `Lexer::flush`
+
 ### Changed
 
 - Items in the `source` module: